+++ tools/actions/convert/test/fixtures/product-opco-home-converted.html
@@ -53,7 +53,7 @@
           </div>
         </div>
       </div>
-      <div class="related-articles opco-home">
+      <div class="articles-list opco-home">
         <div>
           <div></div>
         </div>
