+++ tools/actions/convert/test/fixtures/product-opco-home.html
@@ -482,9 +482,9 @@
 
 </template>
 </grid></div>
-<div class="text aem-GridColumn aem-GridColumn--default--12"><div data-cmp-data-layer="{&#34;related-articles&#34;:{&#34;@type&#34;:&#34;danaher/components/content/marketing/text&#34;,&#34;repo:modifyDate&#34;:&#34;2023-12-05T06:23:37Z&#34;,&#34;xdm:text&#34;:&#34;&lt;p>related-articles&lt;/p>\r\n&#34;}}" id="related-articles" class="cmp-text href-text">
+<div class="text aem-GridColumn aem-GridColumn--default--12"><div data-cmp-data-layer="{&#34;articles-opco-home&#34;:{&#34;@type&#34;:&#34;danaher/components/content/marketing/text&#34;,&#34;repo:modifyDate&#34;:&#34;2023-12-06T11:08:15Z&#34;,&#34;xdm:text&#34;:&#34;&lt;p>articles-opco-home&lt;/p>\r\n&#34;}}" id="articles-opco-home" class="cmp-text href-text">
     
-        <p>related-articles</p>
+  <p>articles-opco-home</p>
 
 </div>
