+++ tools/importer/transformers/index.js
@@ -34,7 +34,7 @@ import libraryHub from './libraryHub.js';
 import sideNav from './sideNav.js';
 import miniTeaser from './miniTeaser.js';
 import productCategoryList from './productCategoryList.js';
-import relatedArticles from './relatedArticles.js';
+import opcoHomeArticlesList from './opcoHomeArticlesList.js';
 import socialFeeds from './socialFeeds.js';
 
 // eslint-disable-next-line import/prefer-default-export
@@ -71,7 +71,7 @@ export const transformers = [
   miniTeaser,
   columns,
   productCategoryList,
-  relatedArticles,
+  opcoHomeArticlesList,
   socialFeeds,
 ];
